+++ client/src/containers/CrossOrder/useHooks.ts
@@ -1,6 +1,6 @@
 import {useEffect} from 'react';
 import {shallowEqual, useDispatch, useSelector} from 'react-redux';
-import {AppState} from 'redux/models/state';
+import {AppState} from 'redux/modules/state';
 import {orderCrossedOnce, postMarketCrossOrder} from 'redux/modules/cross/crossModule';
 import {hasCrossedOnceSelector, hasCrossedSecondTimeSelector, websocketCrossPriceSelector} from 'redux/selectors';
