+++ client/src/redux/store/index.ts
@@ -1,9 +1,9 @@
 import {combineReducers} from 'redux';
 import {configureStore, getDefaultMiddleware} from '@reduxjs/toolkit';
 import reduxWebsocket from '@giantmachines/redux-websocket';
-import notificationMiddleware from '../middlewares/notification';
+import notificationMiddleware from '../middlewares/toast-notification';
 import {API, APIType} from 'redux/api/api';
-import {AppState} from 'redux/models/state';
+import {AppState} from 'redux/modules/state';
 
 import {previewReducer as preview} from '../modules/preview/previewModule';
 import {websocketReducer as websocket} from '../modules/websocket/websocketModule';
