+++ client/src/tests/influnt.ts
@@ -1,6 +1,5 @@
-import {configureInflunt, Inspector, spyModule} from 'influnt';
+import {configureInflunt, spyModule} from 'influnt';
 import {toast} from 'react-toastify';
-import {AppState} from 'redux/models/state';
 import {componentContext, withStore} from './drivers';
 import {networkProxy} from './proxy';
 
@@ -18,21 +17,3 @@ export const createRenderer = configureInflunt({
 export const createComponentRenderer = configureInflunt({
   providerHoc: componentContext,
 });
-
-export function storeActions(): Inspector<string[], Parameters<typeof withStore>[number]> {
-  return ({extraArgs}) => extraArgs.getActions().map(({type}) => type);
-}
-
-export function getState<K extends keyof AppState>(
-  moduleKey: K,
-  key?: keyof AppState[K],
-): Inspector<any, Parameters<typeof withStore>[number]> {
-  return ({extraArgs}) => {
-    const module = extraArgs.getState()[moduleKey];
-    return key ? module[key] : module;
-  };
-}
-
-export function classNameOf(testID: string): Inspector<string | undefined> {
-  return ({locateAll}) => locateAll(testID).className;
-}
